refactor: update folders
