+++ src/react/shared/shared-divider/index.tsx
