+++ src/react/shared/shared-divider/styles.css
