+++ src/react/shared/shared-icon/index.tsx
