+++ src/react/shared/shared-icon/styles.css
