+++ src/react/shared/shared-icon/types.ts
