+++ src/react/shared/shared-icon/utils.tsx
