+++ src/react/shared/shared-stack/index.tsx
