+++ src/react/shared/shared-switch/index.tsx
