+++ src/react/shared/shared-switch/styles.css
