+++ src/react/shared/shared-text/index.tsx
