+++ src/react/table-app/new-row-button/index.tsx
@@ -1,4 +1,4 @@
-import { Button } from "../../shared/button";
+import { Button } from "../../shared/shared-button";
 
 import "./styles.css";
