+++ src/react/table-app/option-bar/search-bar/index.tsx
@@ -1,11 +1,11 @@
-import { Button } from "src/react/shared/button";
-import Icon from "src/react/shared/icon";
-import Stack from "src/react/shared/stack";
+import { Button } from "src/react/shared/shared-button";
+import Icon from "src/react/shared/shared-icon";
+import Stack from "src/react/shared/shared-stack";
 import { setSearchText, toggleSearchBar } from "src/redux/global/global-slice";
 import { useAppDispatch, useAppSelector } from "src/redux/global/hooks";
 
 import "./styles.css";
-import { IconType } from "src/react/shared/icon/types";
+import { IconType } from "src/react/shared/shared-icon/types";
 
 export default function SearchBar() {
 	const { searchText, isSearchBarVisible } = useAppSelector(
