+++ src/react/table-app/tag-cell-edit/component/SelectableTag/index.tsx
@@ -4,10 +4,10 @@ import { MenuLevel, MenuPosition } from "src/shared/menu/types";
 import { useMenu } from "src/shared/menu/hooks";
 
 import "./styles.css";
-import { Button } from "src/react/shared/button";
-import Icon from "src/react/shared/icon";
-import Tag from "src/react/shared/tag";
-import { IconType } from "src/react/shared/icon/types";
+import { Button } from "src/react/shared/shared-button";
+import Icon from "src/react/shared/shared-icon";
+import Tag from "src/react/shared/shared-tag";
+import { IconType } from "src/react/shared/shared-icon/types";
 import { Color } from "src/shared/types";
 import { shiftMenuIntoViewContent } from "src/shared/menu/utils";
