+++ src/react/table-app/tag-cell/index.tsx
@@ -1,6 +1,6 @@
 import { Color } from "src/shared/types";
 import { useOverflowClassName } from "src/shared/spacing/hooks";
-import Tag from "../../shared/tag";
+import Tag from "../../shared/shared-tag";
 
 import "./styles.css";
 interface Props {
